requirements.txt
# ============================================
# CV Algorithm Hub - Python Dependencies
# ============================================

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis>=5.0.0
aioredis>=2.0.1

# Message Queue
aio-pika>=9.3.0
celery>=5.3.0

# Computer Vision & ML
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.8.0
numpy>=1.26.0
pillow>=10.0.0
scikit-image>=0.21.0

# GPU Support (optional)
nvidia-cuda-runtime-cu11>=11.8.0
nvidia-cudnn-cu11>=8.6.0

# Monitoring & Metrics
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0
psutil>=5.9.0

# Logging
structlog>=23.2.0
python-json-logger>=2.0.7

# Configuration
pyyaml>=6.0.1

# HTTP Client
httpx>=0.25.0
aiofiles>=23.2.1

# Utilities
python-multipart>=0.0.6
python-dateutil>=2.8.2
orjson>=3.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0
ruff>=0.1.0
